Port to the TC client on Dalamud API 13
The Traditional Chinese service runs a 7.3-based client on Dalamud 13, two API
levels below upstream.

- Dalamud.NET.Sdk 15 -> 13, which moves the target to net9.0
- Replace the C# 14 constructs that compiler cannot parse: an `extension` block,
  `field` in property accessors, and .NET 10's Enumerable.Shuffle
- Undo v14/v15 API use: IObjectTable.LocalPlayer, IDutyStateEventArgs,
  ISigScanner's namespace, and uint territory-change events
- Bridge the gaps in the API13 ClientStructs snapshot: Character relation and
  weapon flag bits, SpawnObjectPacket, ModelContainer 0x21/0x22, VfxObject's
  member functions and SomeFlags, EnemyListNumberArray's enemy count, Lockon's
  icon column, ILayoutInstance.IsActive
- Drop Scenarios/Umad. Dancing Mad is 7.51 content, so none of its rows, actions
  or VFX exist in a 7.3 client

Every [Signature] is now Fallible, and SignatureReport writes what did and did
not resolve to the log at startup: a miss used to abort the whole plugin load
without saying which one. GetEventObjectByIndex needed attention beyond that —
its four-byte upstream pattern matches three times on this binary and the first
hit lands one byte into an unrelated instruction, so it now covers the whole
function. EventFramework.ProcessDirectorUpdate stays unresolved; its call site
logs and skips rather than dereferencing null.

UI text is Traditional Chinese. Mechanic and form names stay English, which is
how people refer to them. Duty names are read from the client's own
ContentFinderCondition sheet instead of being hardcoded, so they match whatever
the running client calls them. Interactive widgets carry explicit ##ids so their
identity does not move with the translation.

+++ AnoMech/Core/Map/SimHiddenObject.cs
@@ -37,7 +37,7 @@ public sealed unsafe class SimHiddenObject : ISimObject
             if (go.BaseId != baseId) continue;
             var obj = (GameObject*)go.Address;
             obj->DisableDraw();
-            obj->RenderFlags |= VisibilityFlags.Model | VisibilityFlags.Nameplate;
+            obj->RenderFlags |= (int)(VisibilityFlags.Model | VisibilityFlags.Nameplate);
             return new SimHiddenObject(go.ObjectIndex, baseId);
         }
         return null;
@@ -53,7 +53,7 @@ public sealed unsafe class SimHiddenObject : ISimObject
         var obj = Lookup();
         if (obj != null)
         {
-            obj->RenderFlags &= ~(VisibilityFlags.Model | VisibilityFlags.Nameplate);
+            obj->RenderFlags &= ~(int)(VisibilityFlags.Model | VisibilityFlags.Nameplate);
             obj->EnableDraw();
         }
         hidden = false;
